+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - PULP SDK/`librtio`: Add support for floats in `printf`.
+- `pulp` top-level HW module: Add `jtag_tdo_en_o` output pin connected to JTAG debug module.
 
 ### Changed
 - Prefix types in `axi2mem`, `axi_xbar`, and `axi_cut_intf` to prevent type collisions in some